refactor: simplify tool call UI rendering
